Breakup repositories module into multiple modules
