+++ package/kedro_viz/data_access/managers.py
@@ -1,5 +1,6 @@
 """`kedro_viz.data_access.managers` defines data access managers."""
 import logging
+
 # pylint: disable=too-many-instance-attributes
 from collections import defaultdict
 from typing import Dict, List, Set, Union
