+++ package/tests/test_models/test_graph/test_graph_nodes.py
@@ -95,7 +95,6 @@ class TestGraphNodeCreation:
         assert task_node.pipelines == set()
         assert task_node.modular_pipelines == expected_modular_pipelines
 
-    
     def test_task_node_full_name_when_no_given_name(self):
         kedro_node = node(
             identity,
@@ -105,7 +104,6 @@ class TestGraphNodeCreation:
         )
         task_node = GraphNode.create_task_node(kedro_node)
         assert task_node.full_name == "identity"
-
 
     @pytest.mark.parametrize(
         "dataset_name,pretty_name,expected_modular_pipelines",
